Core/requirements.txt
langchain-text-splitters 
langchain-community 
langgraph
sentence-transformers
FlagEmbedding
langchain-google-genai
langchain_huggingface
langchain-qdrant
fastembed
pymongo
fastapi 
uvicorn 
streamlit
